backend/app/api/fundamental_analysis.py
"""
基本面分析
宏观经济因素：如 GDP 增速、利率、通货膨胀率、货币政策等。
行业环境因素：如行业发展阶段、市场规模、竞争格局、政策支持力度等。
公司自身因素：如财务报表（营收、利润、资产负债情况）、商业模式、核心竞争力、管理层能力等。
"""
# app/web/api/analysis.py
from flask_api import APIBlueprint, request, status
from flask import Blueprint
from app.web.api import api_blueprint
from app.web.utils import custom_jsonify
from dependency_injector.wiring import inject, Provide
from app.analysis.fundamental_analyzer import FundamentalAnalyzer
from app.analysis.capital_flow_analyzer import CapitalFlowAnalyzer
from app.analysis._analysis_container import AnalysisContainer
from app.analysis.scenario_predictor import ScenarioPredictor
from app.analysis.stock_qa import StockQA
from app.analysis.risk_monitor import RiskMonitor
from app.analysis.index_industry_analyzer import IndexIndustryAnalyzer
from app.analysis.industry_analyzer import IndustryAnalyzer
from app.analysis.task_manager import TaskManager
from app.analysis.task_manager import TaskStatus
from app.analysis.stock_analyzer import StockAnalyzer
import logging
import traceback
import time
from datetime import datetime

logger = logging.getLogger(__name__)

fundamental_analysis_bp = Blueprint('fundamental_analysis', __name__, url_prefix='/fundamental_analysis')

# 基本面分析
# Fundamental Analysis
@fundamental_analysis_bp.route('/fundamental_analysis', methods=['POST'])
@inject
def api_fundamental_analysis(fundamental_analyzer: FundamentalAnalyzer = Provide[AnalysisContainer.fundamental_analyzer]):
    try:
        data = request.data
        stock_code = data.get('stock_code')
        if not stock_code:
            return {'error': '请提供股票代码'}, status.HTTP_400_BAD_REQUEST
        result = fundamental_analyzer.calculate_fundamental_score(stock_code)
        return custom_jsonify(result)
    except Exception as e:
        logger.error(f"基本面分析出错: {e}", exc_info=True)
        return {'error': str(e)}, status.HTTP_500_INTERNAL_SERVER_ERROR
