backend/pyproject.toml
[project]
name = "stockanal-sys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "akshare>=1.16.98",
    "backtrader>=1.9.78.123",
    "baostock>=0.8.8",
    "bcrypt",
    "beautifulsoup4==4.12.3",
    "chainlit>=2.5.5",
    "chromadb>=1.0.12",
    "dashscope>=1.20.0",
    "dependency-injector>=4.48.1",
    "eodhd>=1.0.32",
    "feedparser>=6.0.11",
    "finnhub-python>=2.4.23",
    "flask-api>=3.1",
    "werkzeug<3.0.0",
    "flask-caching",
    "flask-cors",
    "flask-session>=0.5.0",
    "flask-swagger-ui",
    "gunicorn==20.1.0",
    "html5lib==1.1",
    "ipython>=7.34.0",
    "jsonpath==0.82.2",
    "langchain-anthropic>=0.3.15",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.23",
    "langgraph>=0.4.8",
    "loguru==0.7.2",
    "lxml==4.9.4",
    "markdown>=3.4.0",
    "matplotlib>=3.9.2",
    "numpy>=2.3.0",
    "openai>=1.0.0,<2.0.0",
    "openpyxl==3.1.5",
    "pandas>=2.3.0",
    "parsel>=1.10.0",
    "plotly>=5.0.0",
    "praw>=7.8.1",
    "psutil>=6.1.0",
    "pymongo>=4.0.0",
    "pypandoc>=1.11",
    "pytdx>=1.72",
    "pytest==7.3.1",
    "python-dotenv>=1.0.1",
    "pytz>=2025.2",
    "pyyaml==6.0.2",
    "questionary>=2.1.0",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "scikit-learn>=1.4.0",
    "scipy>=1.14.0",
    "seaborn>=0.13.2",
    "setuptools>=80.9.0",
    "sqlalchemy",
    "statsmodels>=0.14.5",
    "stockstats>=0.6.5",
    "streamlit>=1.28.0",
    "supervisor==4.2.5",
    "tavily-python",
    "tqdm>=4.67.1",
    "tushare>=1.4.21",
    "typing-extensions>=4.14.0",
    "yfinance>=0.2.63",
]


[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
